PSX_OperationStatsImpl: Rework range handling
The current implementation for the ranges did not handle multiple passed
epochs nor epoch wildcards. The check for intersecting ranges added in
6d107905 (PSX_OperationStatsImpl: Assert on intersecting ranges,
2024-02-06) was also broken.

While thinking about the current approach it was recognized that we
actually want to collect the event data from all sweep ranges of all
sweeps from the same equivalence class.

This has now been done. We also support list of epochs and epoch
wildcards.